Fix splits for ftYoshi files
