+++ config/GALE01/symbols.txt
@@ -30287,7 +30287,7 @@ ftNn_Init_804D98E8 = .sdata2:0x804D98E8; // type:object size:0x4 scope:global al
 @304 = .sdata2:0x804D9A20; // type:object size:0x4 scope:local align:4 data:float
 ftYs_Init_804D9A28 = .sdata2:0x804D9A28; // type:object size:0x4 scope:global align:4 data:float
 ftYs_Init_804D9A2C = .sdata2:0x804D9A2C; // type:object size:0x4 scope:global align:4 data:float
-ftYs_Init_804D9A30 = .sdata2:0x804D9A30; // type:object size:0x8 scope:global align:8 data:double
+ftYs_Init_804D9A30 = .sdata2:0x804D9A30; // type:object size:0x8 scope:global align:8 data:8byte
 @175 = .sdata2:0x804D9A38; // type:object size:0x4 scope:local align:4 data:float
 @200 = .sdata2:0x804D9A3C; // type:object size:0x4 scope:local align:4 data:float
 ftYs_Init_804D9A40 = .sdata2:0x804D9A40; // type:object size:0x4 scope:global align:4 data:float
