+++ src/melee/ft/chara/ftYoshi/ftYs_Init.c
@@ -547,76 +547,3 @@ void ftYs_Init_OnKnockbackExit(HSD_GObj* gobj)
     ftAnim_ApplyPartAnim(gobj, 3, 2, 0.0f);
     ftAnim_ApplyPartAnim(gobj, 4, 2, 0.0f);
 }
-
-void ftYs_Init_8012BDA0(Fighter_GObj* gobj)
-{
-    Fighter* fp = GET_FIGHTER(gobj);
-    PAD_STACK(4 * 2);
-    {
-        ftHurtboxInit hurt;
-        PAD_STACK(4 * 4);
-        ftColl_8007B0C0(gobj, Intangible);
-        hurt.bone_idx = fp->ft_data->x8->x11;
-        hurt.height = HurtHeight_Mid;
-        hurt.is_grabbable = true;
-        hurt.a_offset = hurt.b_offset = ftYs_Unk1_803B75C0;
-        hurt.scale = 1;
-        ftColl_HurtboxInit(fp, &fp->hurt_capsules[0], &hurt);
-    }
-}
-
-void ftYs_Init_8012BE3C(HSD_GObj* gobj)
-{
-    /// @todo Some kind of inline here.
-    ftCo_DatAttrs_xBC_t* xBC;
-    Fighter* fp = GET_FIGHTER(gobj);
-    ftParts_80074B0C(gobj, 0, 0);
-    ftColl_8007B0C0(gobj, 0);
-    xBC = &fp->co_attrs.xBC;
-    {
-        Fighter_Part part = ftParts_GetBoneIndex(fp, 4);
-        HSD_JObj* jobj = fp->parts[part].joint;
-        Fighter* fp1 = GET_FIGHTER(gobj);
-        efAsync_Spawn(gobj, &fp1->x60C, 4, 1231, jobj, xBC);
-    }
-}
-
-static void ftYs_Init_8012B8A4_no_inline_2(HSD_GObj* gobj)
-{
-    ftYs_Init_8012B8A4(gobj);
-}
-
-static void ftYs_Init_8012B8A4_no_inline(HSD_GObj* gobj)
-{
-    ftYs_Init_8012B8A4_no_inline_2(gobj);
-}
-
-static void ftYs_Init_8012BECC_sub(HSD_GObj* gobj)
-{
-    Fighter* fp = GET_FIGHTER(gobj);
-    struct {
-        u8 _pad[4];
-        Vec3 v;
-    } s;
-
-    ftColl_8007B0C0(gobj, HurtCapsule_Disabled);
-    fp->mv.ys.guard.xA = 0;
-    fp->mv.ys.guard.x0 = 0;
-    fp->mv.ys.guard.xD = p_ftCommonData->x268;
-    fp->mv.ys.guard.x20 = 0;
-    fp->mv.ys.guard.x1C = 0;
-    s.v.x = s.v.y = s.v.z = 0;
-    HSD_JObjSetTranslate(fp->parts[fp->ft_data->x8->x11].joint, &s.v);
-    ftYs_Init_8012B8A4_no_inline(gobj);
-    ftCo_80091D58(fp);
-    ft_PlaySFX(fp, 0x6E, 0x7F, 0x40);
-}
-
-void ftYs_Init_8012BECC(Fighter_GObj* gobj)
-{
-    PAD_STACK(8);
-    Fighter_ChangeMotionState(gobj, ftYs_MS_GuardOn_0, 0, 0.0f, 1.0f, 0.0f, NULL);
-    ftAnim_8006EBA4(gobj);
-    ftCo_80092450(gobj);
-    ftYs_Init_8012BECC_sub(gobj);
-}
